#147 test: 세부 일정내의 특정 여행지를 제거하는지 테스트
